[refactor] #78 dto에 valid 추가 및 폴더 이름 변경
